# backend/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "URL Shortener Backend API",
        "version": "1.0.0",
        "description": "API documentation for the URL shortening service"
    },
    "servers": [
        {
            "url": "http://localhost:5000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://your-deployed-url/api",
            "description": "Deployed Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "rateLimit": {
                "type": "apiKey",
                "in": "header",
                "name": "X-Rate-Limit",
                "description": "Rate limiting applied (100 requests per 15 minutes)"
            }
        },
        "schemas": {
            "Url": {
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "originalUrl": {
                        "type": "string",
                        "format": "uri"
                    },
                    "shortUrl": {
                        "type": "string",
                        "format": "uri"
                    },
                    "urlCode": {
                        "type": "string"
                    },
                    "qrCode": {
                        "type": "string",
                        "format": "uri"
                    },
                    "clicks": {
                        "type": "number"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastAccessed": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "ipAddress": {
                        "type": "string"
                    },
                    "userAgent": {
                        "type": "string"
                    }
                },
                "required": [
                    "originalUrl",
                    "shortUrl",
                    "urlCode",
                    "qrCode"
                ]
            },
            "UrlHistory": {
                "type": "object",
                "properties": {
                    "originalUrl": {
                        "type": "string",
                        "format": "uri"
                    },
                    "shortUrl": {
                        "type": "string",
                        "format": "uri"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "clicks": {
                        "type": "number"
                    },
                    "qrCode": {
                        "type": "string",
                        "format": "uri"
                    }
                }
            },
            "ShortenRequest": {
                "type": "object",
                "properties": {
                    "originalUrl": {
                        "type": "string",
                        "format": "uri"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    }
                },
                "required": [
                    "originalUrl"
                ]
            }
        }
    },
    "paths": {
        "/shorten": {
            "post": {
                "summary": "Shorten a URL",
                "description": "Creates a shortened URL with an associated QR code",
                "tags": [
                    "URL Management"
                ],
                "security": [
                    {
                        "rateLimit": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ShortenRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL shortened successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Url"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid URL format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/history": {
            "get": {
                "summary": "Get URL shortening history",
                "description": "Retrieves a list of all shortened URLs with basic stats",
                "tags": [
                    "URL Management"
                ],
                "security": [
                    {
                        "rateLimit": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UrlHistory"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{code}": {
            "get": {
                "summary": "Redirect to original URL",
                "description": "Redirects to the original URL and increments click count",
                "tags": [
                    "URL Management"
                ],
                "security": [
                    {
                        "rateLimit": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique URL code"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to original URL"
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stats/{code}": {
            "get": {
                "summary": "Get URL statistics",
                "description": "Retrieves detailed statistics for a specific shortened URL",
                "tags": [
                    "URL Management"
                ],
                "security": [
                    {
                        "rateLimit": []
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The unique URL code"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Url"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "summary": "Health check",
                "description": "Returns the health status of the server",
                "tags": [
                    "System"
                ],
                "responses": {
                    "200": {
                        "description": "Server is healthy",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "uptime": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "URL Management",
            "description": "Endpoints for managing URL shortening and statistics"
        },
        {
            "name": "System",
            "description": "System health and status endpoints"
        }
    ]
}